Removed unreported values from options
Currently this sits in the amp as an array of sensor types that don't support the option as of now. May move to object to include firmware versions where support is added. Possibly we can add functionality to get these options based on what a sensor reported, but would require passing in a sensor address.